Revert "Remove hashes on built filenames (#13567)"
This reverts commit a1054455353fe6c0a27653c4e95abcaafc639d18.
