--- .changeset/small-rocks-grab.md
@@ -1,5 +1,0 @@
----
-"react-router": patch
----
-
-Remove hashes from files in `dist/` for easier usage with `patch-package`
