+++ integration/prefetch-test.ts
@@ -172,7 +172,7 @@ test.describe("prefetch", () => {
           await page.waitForSelector(
             // Look for either Rollup or Rolldown chunks
             [
-              "link[rel='modulepreload'][href^='/assets/index-']",
+              "link[rel='modulepreload'][href^='/assets/chunk-']",
               "link[rel='modulepreload'][href^='/assets/jsx-runtime-']",
             ].join(","),
             { state: "attached" }
@@ -232,7 +232,7 @@ test.describe("prefetch", () => {
           await page.waitForSelector(
             // Look for either Rollup or Rolldown chunks
             [
-              "link[rel='modulepreload'][href^='/assets/index-']",
+              "link[rel='modulepreload'][href^='/assets/chunk-']",
               "link[rel='modulepreload'][href^='/assets/jsx-runtime-']",
             ].join(","),
             { state: "attached" }
@@ -251,7 +251,7 @@ test.describe("prefetch", () => {
           await page.waitForSelector(
             // Look for either Rollup or Rolldown chunks
             [
-              "link[rel='modulepreload'][href^='/assets/index-']",
+              "link[rel='modulepreload'][href^='/assets/chunk-']",
               "link[rel='modulepreload'][href^='/assets/jsx-runtime-']",
             ].join(","),
             { state: "attached" }
@@ -341,7 +341,7 @@ test.describe("prefetch", () => {
           await page.waitForSelector(
             // Look for either Rollup or Rolldown chunks
             [
-              "link[rel='modulepreload'][href^='/assets/index-']",
+              "link[rel='modulepreload'][href^='/assets/chunk-']",
               "link[rel='modulepreload'][href^='/assets/jsx-runtime-']",
             ].join(","),
             { state: "attached" }
@@ -360,7 +360,7 @@ test.describe("prefetch", () => {
           await page.waitForSelector(
             // Look for either Rollup or Rolldown chunks
             [
-              "link[rel='modulepreload'][href^='/assets/index-']",
+              "link[rel='modulepreload'][href^='/assets/chunk-']",
               "link[rel='modulepreload'][href^='/assets/jsx-runtime-']",
             ].join(","),
             { state: "attached" }
