+++ packages/react-router/lib/hooks.tsx
@@ -49,7 +49,7 @@ import {
   resolveTo,
   stripBasename,
 } from "./router/utils";
-import type { SerializeFrom } from "./types/internal-export/route-data";
+import type { SerializeFrom } from "./types/route-data";
 
 /**
   Resolves a URL against the current location.
