+++ packages/react-router/package.json
@@ -39,13 +39,13 @@
     },
     "./internal": {
       "node": {
-        "types": "./dist/development/internal-export.d.ts"
+        "types": "./dist/development/lib/types/internal.d.ts"
       },
       "import": {
-        "types": "./dist/development/internal-export.d.mts"
+        "types": "./dist/development/lib/types/internal.d.mts"
       },
       "default": {
-        "types": "./dist/development/internal-export.d.ts"
+        "types": "./dist/development/lib/types/index.d.ts"
       }
     },
     "./dom": {
